Fix flaky test
Use await_event_for so that batching of ignored presence events doesn't
matter.

This is a little questionable. The desired semantics for await_event_for
called across different room_ids is unclear. Probably saved_events
should be a hash of room_id => array of events, but that's quite fiddly
to set up.