Update sytests to match e2e key upload changes
The e2e endpoints have never been specced, so we should be accessing them on an
/unstable endpoint for now.

The 'deviceId' parameter on e2e key upload is now optional (and where it
exists, it must match that from the login) - so we can no longer just make it
up. Stash the device_id when we register or login, and use it in the e2e keys.

GET /keys/upload has never even been proposed as an API, and has now been
removed from synapse. Test that keys have been claimed using a second call to
/claim instead.