Fix some races in the device-lists tests
both matrix_put_e2e_keys and matrix_set_device_display_name cause a device
"changed" notification, so it was impossible to tell if we'd received both
notifications - which meant that sometimes we would carry on with the test and
the notification would arrive late.

Fix this either by waiting for the changed notification after the first call,
or by dropping the call to matrix_set_device_display_name.

Also factor out the 'sync until the notification appears' code.