Fix the fact that printing PDUs changes them
This is a workaround for a tragic, tragic tale which comes down to the fact
that Perl lacks a proper way to distinguish between a string and a number.

Following the Perl philosophy of "I'm going to guess what you meant and usually
get it wrong", the JSON encoder therefore has to guess whether you're trying to
send '"0"' or '0', which it does by digging in the guts of the interpreter to
see whether the scalar in question was last referred to as a string or a
number.

However, the pretty-printer used by log_if_fail, Data::Dump, turns out to
mangle its inputs just enough to turn numbers into strings. Which means,
hilariously, that if you try to log_if_fail an event, everything breaks.

So... the workaround for this is to wrap things that we want to stay as numbers
with our JSON::number class; we then need to tell the JSON encoder to use
blessed objects' TO_JSON method rather than barfing (hence, `convert_blessed`).

Because why would you ever want to tell numbers and strings apart.