Don't require /sync when testing room create
The room creation test is marked "critical", which means that if it
fails the entire test run is aborted. However, we don't want to abort
the test run if there is a bug in the sync handling, so we split the
test into two.

We test the fact that the room create comes down sync stream separately.