Don't set PYTHONPATH when running synapse (#560)
Setting PYTHONPATH runs the risk of picking up random files from elsewhere in
the synapse readme (for instance, we might accidentally import something from
the 'tests' hierarchy).

Our recommendation for installing synapse (and what the sytest docker image
does) is to install it into the virtualenv. This means that there is no need to
set PYTHONPATH any more.