Fix apt incantation in dockerfiles
We need to combine the update and install into one dockerfile command. If we do
not, docker can choose to use a cached version of (base image + `apt-get
update`), which is out-of-date, so that the install fails.

Combining them into one RUN command means that docker will always run the
update if it's going to run the install.