Add default responder to HTTP server
This lets one add a default responder to a path that gets called if no
handlers match that path. This is useful when registering paths for push
etc where you want requests to be handled correctly rather than dropped
on the floor if no test is currently awaiting for it.