uri_escape key_id on /key/v2/server endpoint (#635)
It was possible for `$key_id` to have a `/` in it, which caused some homeservers to not path-match correctly (and rightly so). Make sure we URL encode before sending to the server.

(This worked on Synapse because the keyID in this endpoint is actually deprecated and Synapse ignores it entirely).