Bind the mocked federation server to $BIND_HOST
`$BIND_HOST` and `$sock->sockhostname` don't always hold the same value (i.e. if the system has different hostnames for the localhost). Since we need the mocked federation server's hostname to match its certificate's common name, and the certificate is generated for `$BIND_HOST`, this changes the server's init so that it uses `$BIND_HOST` as its hostname as well.