Have the in-built identity server support v2 (#689)
Add all of the v2 identity server endpoints to Sytest's built-in identity server. This is independent of any of the Synapse changes currently in-flight.

Additionally, change the identity server tests to use these v2 endpoints.

This essentially just involves switching the endpoints to v2 and including an `id_access_token` parameter at the top-level of the JSON, or in a query parameter. I tried to reuse functions wherever possible.

The `id_access_token` is hard-coded, as there isn't much point retrieving it via OpenID as we're using a built-in Sytest identity server, which we don't need to test. Sydent's OpenID functionality is handled by matrix-is-tester.