Fix flakiness due to create_room_synced (#702)
When we create a room, it's possible for a subsequent /sync to receive subset of
the intial events in that room. This was making some tests [1] flaky when
running in worker mode.

AFAICT there's nothing in the spec which forbids this behaviour. By way of a
workaround, we send a test message in the new room and wait for it to turn up
over /sync.

However, a corollary is that we can't rely on the /sync to contain all the
events in the room. Solution to that is to sync again in tests that actually
care about the sync body.

[1]: notably 'A message sent after an initial sync appears in the timeline of
an incremental sync.'