do requestToken before adding a 3pid (#706)
For enhanced realism (TM), hit the requestToken endpoint before attempting to
add a 3pid to a user account.

Currently, we can get away without hitting requestToken, since it just
delegates to the IS, which we stub out. However, MSC2078 changes this so that a
homeserver is free to handle requestToken itself, which makes it a required
part of the flow.

This PR therefore lays the groundwork for future work which will handle the
homeserver implementing requestToken itself.