Add Perl results.tap parser (#714)
The aim of this is to replace the [format_tap.py](https://github.com/matrix-org/synapse/blob/develop/.buildkite/format_tap.py) file that's required to show Buildkite annotations for failed Sytest tests.

This perl script generates a line-for-line same output as the python version. This is useful so that we don't need to include a python binary for Dendrite's sytest environment.