Don't add outliers to forward extremities.
The inbuilt federation library blindly added all events as foreward
extremities. Normally this would be fine as adding all state events in
the correct order removed all superfluous extremities, however if not
done in the right order, or there was a gap, then new events would get
sent out with a random assortment of prev events. This led to some
events being rejected due to the resolved state being not what was
expected.