Fix race in 'Outbound federation can backfill events'
This test will issue a `/join` for a remote room alias then attempt
to backpaginate in the room. The way the backpagination works
presents a race condition. Previously it:
 - Sent the `/join`, wait for 200 OK.
 - Hit `/sync` and remember `next_batch`
 - Hit `/messages` with `next_batch`.
 - If not enough messages, hit `/messages` again with `end` token from
   `/messages`.

In Dendrite, it's possible that `/sync` has not seen the join membership
event yet, resulting in a `next_batch` that represents the state *before*
the membership event. Backpagination then gets messages prior to the join
event, when the test expects it to include the join event. The fix is to
wait for the `m.room.member` event to come down `/sync` and THEN use the
`next_batch` token.