Use matrix_invite_user_to_room_synced on remote invite tests
Dendrite will return 200 OK to `/invite` after the invite event has
been enqueued to the `federationsender` but before it has made its way
through all the components. Previously, this test would immediately hit
`GET /state/m.room.member/userID` which would sometimes 404 as the state
had not reached that component. This PR modifies this behaviour to wait
for the event to come down `/sync`, which serves as a good proxy for
having reached all dendrite components, then do the GET.