Wait for synapse to start by using sd_notify.
Currently SyTest waits for synapse process to fully start by repeatedly
checking if it can connect to the HTTP socket. This won't work for
starting worker processest that do not listen on any HTTP socket.

Instead, we use the `sd_notify` mechanism. This works by SyTest binding
to a unix socket, passing the socket address to Synapse via env, and
waiting to receive a `READY=1` notification on the socket.