Deflake "presence in initialSync" test
this seems to have looped waiting for presence to turn up ever since it was
first written, but as far as I can tell, the loop has never done anything,
because where we don't know presence, we always make up an "offline" status.

Anyway, this seems to make it more reliable.