Bump the version of Future we're depending on
For the second time this week (#972 and #973), I've found myself wondering why
a test was failing and discovered that it is due to it relying on new
functionality in Future 0.45. I guess we may as well bump the dep as try to
keep our tests backwards-compatible, particularly given the CI uses the latest
version.