Update soft-fail test 1 to use `make_event_refs`
Rather than manually building the (event_id, event_hash) prev_event structures,
use `$room->make_event_refs` to do this, which means it will also work on later
room verions.

This means we need access to the events themselves rather than just the IDs, so
needs a bit of tweaking for that.