Set auth_events explicitly
When we create an event with prev_events which put that event on a historical
branch, we also need to use the auth_events corresponding to that point on the
DAG. Otherwise, the server can end up pulling in the state from the main branch
anyway, and rejecting, rather than soft-failing, the relevant events.

(Exactly whether the state gets pulled in from the auth_events depends on the
state resolution algorithm in use, hence on the room version).