Make "Notifications can be viewed with GET /notifications" sync before sending receipts.
In Dendrite, things are run on asynchronous Kafka queue, and it may
happen that a receipt is discarded unless the event ID actually
exists. In practice, no client would send a receipt for an event that
it hasn't seen in the `/sync` stream, so making this synchronous seems
like it would emulate the real world better.

There is still a race: just because Dendrite's Sync API has seen the
sent message doesn't mean Push server has, but in practice this seems
enough to make the test pass with https://github.com/matrix-org/dendrite/pull/1842.