Fix scripts/dendrite_sytest.sh to show are-we-synapse-yet in the console, and chmod logs.
Some earlier changed made a redirect of the output to file, and
(probably inadvertently) removed the console output. This is a very
useful piece of information.

Since Sytests run as root, it's practical if copied-out logs are
chmod:ed to be accessible by whoever started the Docker job. For test
logs, making them world readable feels OK.