Mark `testing` image as experimental
This marks the `testing` image as experimental and allows the builds of the stable images to continue even if the experimental ones fail.

It's nice to know when something in `testing` will break CI for a future Debian version, but it probably shouldn't stop us from successfully building existing images based on the stable distributions.

(Related to #1114)