Resolve room alias in `matrix_create_room`
The spec does not require `/createRoom` request to return room aliases
(https://spec.matrix.org/v1.2/client-server-api/#post_matrixclientv3createroom),
but `matrix_create_room` tries to read it from the response. This field
is deprecated and does not appear on the spec.

For those implementations which are not returning the alias, we will
issue a new request to load room's aliases in order to make this
function compatible.

Signed-off-by: Guillem Nieto <gnieto.talo@gmail.com>