Respect poetry lockfile in Synapse container, if present (#1220)
Install pinned dependencies using poetry if a `poetry.lock` file is
present in the Synapse source directory, otherwise fall back to pip.

The Synapse container installs Synapse dependencies twice, once at
Docker build time and once at test time. The former installation is
intended to speed up the latter. To make this poetry-compatible, we
have to convince poetry to use the same virtual env both times. By
using the `virtualenvs.in-project` config option, we can get poetry to
use a virtual env at `./.venv`. The source directory is going to be a
mounted volume and writing into it is impolite, so we have to make a
copy of it. For simplicity, we ditch the tarball for non-poetry
installs and always make a copy.

To make offline mode work, we have to make a cached copy of a
`poetry-core` wheel available to pip. For simplicity, we do the same
thing for `setuptools` and `wheel` for pre-poetry Synapses.

Once Synapse develop has been converted to poetry, the TODO in
`docker/synapse.Dockerfile` can be resolved.

Signed-off-by: Sean Quah <seanq@element.io>