Use speced redaction endpoints
Sytest uses an endpoint for redaction which uses POST (instead
of PUT) and does not contain a transaction id. I've been reviewing the
latest spec version and I could only find the PUT version
(https://spec.matrix.org/v1.2/client-server-api/#put_matrixclientv3roomsroomidsendeventtypetxnid).

I've changed the endpoints to use the speced endpoint (both synapse and
dendrite seems to support both). I'm not certain about if this is an
issue on the spec or in sytest: Feel free to close this PR if the
previous behaviour is the expected one.

Signed-off-by: Guillem Nieto <gnieto.talo@gmail.com>