Attempt to fix synapse-sytest docker images (#1384)
BBind to 127.0.0.1 when using docker

I'm not sure why this is a problem, but emprically it is now required
for me to run it locally.

The exact thing that needs to change is that HaProxy server addresses
need to be changed. I assume its something like when HaProxy starts
`localhost` can't be resolved.