Stop using 'database' and start using 'dbname' to describe the database name to connect to in Postgres
For Synapse, these pass through and match expected configuration expectations from the manual.

For Dendrite, the generated database config is parsed and turned into a DSN that is used to
generate its main config yaml. All that was needed is to update where the value comes from.

This brings the named parameters into line with libpq.