Update configurations to recognize `dbname` instead of `database` (#1390)
* Stop using 'database' and start using 'dbname' to describe the database name to connect to in Postgres

For Synapse, these pass through and match expected configuration expectations from the manual.

For Dendrite, the generated database config is parsed and turned into a DSN that is used to
generate its main config yaml. All that was needed is to update where the value comes from.

This brings the named parameters into line with libpq.

* Stop removing and re-adding the `name` attribute from the database config

This only seemed to be needed for Synapse as Dendrite used another method and was not relevant