Deflake 'uploading signed devices gets propagated over federation'
This test started to fail after implementing
https://github.com/element-hq/synapse/pull/18899. The failure
appeared to be due to the test not waiting for signatures to
propagate over federation. The  loop
would exit as soon as it saw a 'signatures' key. But the
value was an empty dict.

A few moments later, the dict would be populated with the key
we were waiting for. But while that was being sent over federation,
the test would fail and exit abruptly.

This commit changes the  loop to actually
check for the signature we're waiting for, instead of exiting
upon seeing the 'signatures' key. This seems like it would be
less flaky in general, and prevents the test from failing in
this case.